Feature MapUI and Updated textures

+++ Tower Defense/Scenes/Components/Projectile/ProjectileInfo.cs
@@ -5,5 +5,5 @@ using System;
 public partial class ProjectileInfo : Resource
 {
 	[Export] public float Speed;
-	[Export] public float Damage;
+	[Export] public int Damage;
 }
